Changes to how a fasta file is parsed into peptides.  Instead of generating all peptides from the file and storing them in memory, they are generated one at a time as requested through the peptide iterator. It is now possible to build a non-enzyme specific index of yeast without running out of memory.
Sixteen smoke tests changed.  Thirteen of those require that decoys be
scored to get enough scores to estimate p-values.  Different decoys are
generated, so the distribution of scores changes slightly.  For one
spectrum, the change of p-value is small enough that I think it is
a floating point difference.  I can't account for the change for the
last two spectra.  There were also three cases where the current version
produces one fewer sequence for a spectrum.  In all three cases, the
old version was producing the same sequence twice.

The compute_qvalues test is based on the above search_pvalues test.
Since the p-values were slightly different, the q-values got shifted
a little. The difference between the old value and the new value is no
more than 0.03.  The decoy_qval test had changes for similar reasons.